scripts/wordcloud/requirements.txt
wordcloud==1.7.0
requests==2.22.0
requests-oauthlib==1.2.0
Pillow==8.3.2
numpy==1.19.0
Janome==0.3.10
emoji==0.5.4
tweepy==3.9.0
beautifulsoup4==4.9.1
reppy==0.4.14
markovify==0.8.3
retrying==1.3.3
